Correct code.

+++ pkg/add/stub.go
@@ -3,4 +3,4 @@
 package add
 
 // Add adds x and y.
-func Add(x uint64, y uint64) uint32
+func Add(x uint64, y uint64) uint64
